+++ lib/lib.rs
@@ -33,7 +33,7 @@ let mut server: Server = Server::new("localhost", 8080);
 
 // Add a route
 server.route(Method::GET, "/greet/{name}", |req| {
-  let name = req.path_pram("name").unwrap();
+  let name = req.path_param("name").unwrap();
 
   Response::new()
     .text(format!("Hello, {}", name))